+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
 
 
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
 
             <intent-filter>
