+++ app/src/main/java/com/example/logintask1/ui/MainActivity.kt
@@ -1,9 +1,9 @@
-package com.example.logintask1
+package com.example.logintask1.ui
 
 import android.os.Bundle
-import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import com.example.logintask1.R
 import com.example.logintask1.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
