+++ app/src/main/java/com/example/logintask1/ui/auth/signin/SigninFragment.kt
@@ -1,25 +1,24 @@
-package com.example.logintask1.signupPackage
+package com.example.logintask1.ui.auth.signin
 
 import android.os.Bundle
-
 import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.example.logintask1.R
-import com.example.logintask1.databinding.FragmentSignupBinding
+import com.example.logintask1.databinding.FragmentSigninBinding
 
 
-
-class SignupFragment: Fragment(R.layout.fragment_signup) {
-    private lateinit var binding: FragmentSignupBinding
-    private val model: SignupViewModel by viewModels()
+class SigninFragment : Fragment(R.layout.fragment_signin) {
+    private lateinit var binding: FragmentSigninBinding
+    private val model: SigninViewModel by viewModels()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding = FragmentSignupBinding.bind(view)
+
+        binding = FragmentSigninBinding.bind(view)
         binding.viewModel = model
         binding.lifecycleOwner = this
 
     }
-
 }
+
