+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="com.example.logintask1.signinPackage.SigninFragment"
+        android:name="com.example.logintask1.ui.auth.signin.SigninFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toTopOf="@+id/constraintLayout2"
