+++ app/src/main/res/layout/fragment_signup.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.example.logintask1.signupPackage.SignupViewModel" />
+            type="com.example.logintask1.ui.auth.signup.SignupViewModel" />
     </data>
 
 <androidx.constraintlayout.widget.ConstraintLayout
